+++ grunt/watch.js
@@ -1,6 +1,6 @@
 module.exports = {
-	dist : {
-		files : [  'src/**/*.js' ],
-		tasks : [ "newer:concat", "newer:uglify" ]
-	}		
+	dist: {
+		files: ["src/**/*.js"],
+		tasks: ["newer:concat", "newer:terser"],
+	},
 };